fix some XXX
