+++ mne/channels/tests/test_montage.py
@@ -29,7 +29,6 @@ from mne.channels import compute_dev_head_t
 
 from mne.channels._dig_montage_utils import _transform_to_head_call
 from mne.channels._dig_montage_utils import _fix_data_fiducials
-from mne.channels._dig_montage_utils import _get_fid_coords
 from mne.utils import (_TempDir, run_tests_if_main, assert_dig_allclose,
                        object_diff, Bunch)
 from mne.bem import _fit_sphere
@@ -37,6 +36,7 @@ from mne.transforms import apply_trans, get_ras_to_neuromag_trans
 from mne.io.constants import FIFF
 from mne._digitization import Digitization
 from mne._digitization._utils import _read_dig_points, _format_dig_points
+from mne._digitization._utils import _get_fid_coords
 from mne._digitization.base import _get_dig_eeg, _count_points_by_type
 
 from mne.viz._3d import _fiducial_coords
